refactor: AdminController 패키지 계층 적용
- 쓸모없는 import 제거
