+++ src/main/java/roomescape/admin/presentation/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.presentation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
