+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,7 +1,6 @@
 package roomescape.reservation.dto;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 
 public record ReservationRequest(String name, LocalDate date, Long timeId) {
 }
